refactor(config): improve code style and configuration handling

+++ database/migrations/2025_06_19_072729_create_provinces_table.php
@@ -8,22 +8,26 @@ return new class extends Migration
 {
     public function up()
     {
-        $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'provinsi' : 'provinces');
-        $code = config('indonesia.pattern') === 'ID' ? 'kode' : 'code';
-        $name = config('indonesia.pattern') === 'ID' ? 'nama' : 'name';
+        if (config('indonesia.data_location.province')) {
+            $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'provinsi' : 'provinces');
+            $code = config('indonesia.pattern') === 'ID' ? 'kode' : 'code';
+            $name = config('indonesia.pattern') === 'ID' ? 'nama' : 'name';
 
-        Schema::create($tableName, function (Blueprint $table) use ($code, $name) {
-            $table->bigIncrements('id');
-            $table->char($code, 2)->unique();
-            $table->string($name, 255);
-            $table->text('meta')->nullable();
-            $table->timestamps();
-        });
+            Schema::create($tableName, function (Blueprint $table) use ($code, $name) {
+                $table->bigIncrements('id');
+                $table->char($code, 2)->unique();
+                $table->string($name, 255);
+                $table->text('meta')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
     {
-        $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'provinsi' : 'provinces');
-        Schema::drop($tableName);
+        if (config('indonesia.data_location.province')) {
+            $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'provinsi' : 'provinces');
+            Schema::drop($tableName);
+        }
     }
 };
