+++ database/migrations/2025_06_19_072819_create_villages_table.php
@@ -8,32 +8,44 @@ return new class extends Migration
 {
     public function up()
     {
-        $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kelurahan' : 'villages');
+        if (config('indonesia.data_location.village')) {
+            $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kelurahan' : 'villages');
+            $code = config('indonesia.pattern') === 'ID' ? 'kode' : 'code';
+            $name = config('indonesia.pattern') === 'ID' ? 'nama' : 'name';
 
-        $code = config('indonesia.pattern') === 'ID' ? 'kode' : 'code';
-        $name = config('indonesia.pattern') === 'ID' ? 'nama' : 'name';
+            if (config('indonesia.data_location.district')) {
+                $districtTable = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kecamatan' : 'districts');
+                $districtCode = config('indonesia.pattern') === 'ID' ? 'kode_kecamatan' : 'district_code';
+            } else {
+                $districtTable = null;
+                $districtCode = null;
+            }
 
-        $districtTable = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kecamatan' : 'districts');
-        $districtCode = config('indonesia.pattern') === 'ID' ? 'kode_kecamatan' : 'district_code';
+            Schema::create($tableName, function (Blueprint $table) use ($code, $name, $districtCode, $districtTable) {
+                $table->bigIncrements('id');
+                $table->char($code, 10)->unique();
+                if (config('indonesia.data_location.district')) {
+                    $table->char($districtCode, 7);
+                }
+                $table->string($name, 255);
+                $table->text('meta')->nullable();
+                $table->timestamps();
 
-        Schema::create($tableName, function (Blueprint $table) use ($code, $name, $districtCode, $districtTable) {
-            $table->bigIncrements('id');
-            $table->char($code, 10)->unique();
-            $table->char($districtCode, 7);
-            $table->string($name, 255);
-            $table->text('meta')->nullable();
-            $table->timestamps();
-
-            $table->foreign($districtCode)
-                ->references($code)
-                ->on($districtTable)
-                ->onUpdate('cascade')->onDelete('restrict');
-        });
+                if (config('indonesia.data_location.district')) {
+                    $table->foreign($districtCode)
+                        ->references($code)
+                        ->on($districtTable)
+                        ->onUpdate('cascade')->onDelete('restrict');
+                }
+            });
+        }
     }
 
     public function down()
     {
-        $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kelurahan' : 'villages');
-        Schema::drop($tableName);
+        if (config('indonesia.data_location.village')) {
+            $tableName = config('indonesia.table_prefix') . (config('indonesia.pattern') === 'ID' ? 'kelurahan' : 'villages');
+            Schema::drop($tableName);
+        }
     }
 };
